+++ AddOns/TargetNameplateIndicator/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.38
+- Add Target of Target indicator
+- Refactor checking of other indicators to priority-based system
+
+## 1.37
+- Bump TOC Interface version to 9.0.5
+
 ## 1.36
 - Bump TOC Interface version to 9.0.1